requirements_tests.txt
pytest
pytest-cov
coverage
flake8
pylint
mypy
setuptools
wheel

gitpython
semver
